requirements.txt
# Core LLM & Framework
langchain>=0.3.0
langchain-community>=0.3.0
langchain-ollama>=0.2.0
langchain-groq>=0.2.0

# UI
streamlit>=1.40.0

# Data
pandas>=2.2.0
openpyxl>=3.1.0
pypdf>=4.0.0
tabulate>=0.9.0

# Charts & Export
plotly>=5.24.0
kaleido>=0.2.1
python-docx>=1.1.0
reportlab>=4.0.0

# Image & OCR
Pillow>=10.0.0
pytesseract>=0.3.10

# Language Detection
langdetect>=1.0.9

# Config
python-dotenv>=1.0.0

# OpenAI
openai>=1.30.0
langchain-openai>=0.1.0

# Anthropic
anthropic>=0.30.0
langchain-anthropic>=0.1.0

# Google Gemini
google-generativeai>=0.7.0
langchain-google-genai>=1.0.0

# Phase 2
chromadb>=0.5.0
sentence-transformers>=3.0.0
